requirements.txt
torch>=1.9.0
regex>=2021.8.3
requests>=2.26.0
pyyaml>=5.4.1
numpy>=1.20.0
